[Rename package: 'mapDownload' -> 'map.download' + Merge conflict fixes] - Rename package: 'games.strategy.engine.framework.mapDownload' to: 'games.strategy.engine.framework.map.download' - Restore LocalSystemChecker, it was recently removed incorrectly during merge conflict resolution. - Rename method variable from 'e' to 'event'. Not strictly necessary, but will help avoid the same variable name from being used down on line 130 with 'SwingAction.of( e -> ...'
