+++ src/games/strategy/engine/framework/map/download/DownloadFileParser.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 import static com.google.common.base.Preconditions.checkState;
