+++ src/games/strategy/engine/framework/map/download/MapDownloadController.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 import java.awt.Frame;
 import java.io.File;
