+++ src/games/strategy/engine/framework/map/download/MapDownloadHelpPanel.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 import javax.swing.JLabel;
 import javax.swing.JPanel;
