+++ src/games/strategy/engine/framework/map/download/MapDownloadProgressPanel.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 import java.awt.GridLayout;
 import java.util.List;
