+++ test/games/strategy/engine/framework/map/download/DownloadFileDescriptionTest.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -9,6 +9,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import games.strategy.engine.ClientFileSystemHelper;
+import games.strategy.engine.framework.map.download.DownloadFileDescription;
 import games.strategy.util.Version;
 
 public class DownloadFileDescriptionTest {
