+++ test/games/strategy/engine/framework/map/download/FileDownloadTest.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.mapDownload;
+package games.strategy.engine.framework.map.download;
 
 
 import static org.hamcrest.Matchers.is;
@@ -12,7 +12,10 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import games.strategy.engine.framework.mapDownload.DownloadFile.DownloadState;
+import games.strategy.engine.framework.map.download.DownloadFile;
+import games.strategy.engine.framework.map.download.DownloadFileDescription;
+import games.strategy.engine.framework.map.download.MapDownloadStrategy;
+import games.strategy.engine.framework.map.download.DownloadFile.DownloadState;
 
 
 @RunWith(MockitoJUnitRunner.class)
